+++ jwst/regtest/sdp_pools_source.py
@@ -2,41 +2,43 @@
 # Setup a base class and instantiate it in order to provide the
 # SDP pool lists for test parametrization.
 # #############################################################
-from jwst.tests.base_classes import BaseJWSTTest
+import os
+
+from jwst.regtest.regtestdata import RegtestData
 
 
-class SDPPoolsSource(BaseJWSTTest):
+class SDPPoolsSource(RegtestData):
     """Retreive the SDP pools from the test data source
 
-    These tests are very much tied to using `pytest` and the `ci-watson` plugin.
-    In particular, there are references to `pytest.config` that only exist when
-    running under pytest. Such references are stubbed out with best defaults used.
+    This class defines a common relationship between test collection, via
+    `jwst.regtest.conftest.pool_path_fixture` and the test using said fixture,
+    `jwst.regtest.test_associations_sdp_pools`.
     """
 
-    # The following will be set during both plugin instantiation
-    # of `BaseJWSTTest`, or explicitly during dynamic parametrization
-    # The empty string defaults indicate that data source roots have
-    # been defined.
-    inputs_root = ''
-    results_root = ''
-
-    input_loc = 'associations'
-    test_dir = 'sdp'
+    # Setup some defaults.
+    test_dir = 'associations/sdp'
     ref_loc = [test_dir, 'truth']
 
+    _inputs_root = ''
     _pool_paths = None
-    _truth_paths = None
+    _results_root = ''
+
+    def __init__(self, okify_op='folder_copy', **kwargs):
+        super().__init__(okify_op=okify_op, **kwargs)
 
     @property
     def pool_paths(self):
         """Get the association pools"""
         if self._pool_paths is None:
-            self._pool_paths = self.data_glob(self.test_dir, 'pools', glob='*.csv')
+            self._pool_paths = self.data_glob(self.test_dir + '/pools', glob='*.csv')
         return self._pool_paths
 
-    @property
-    def truth_paths(self):
+    def truth_paths(self, pool):
         """Get the truth associations"""
-        if self._truth_paths is None:
-            self._truth_paths = self.data_glob(*self.ref_loc, glob='*.json')
-        return self._truth_paths
+        paths = []
+        truth_pool_path = '/'.join(self.ref_loc) + '/' + pool
+        for path in self.data_glob(truth_pool_path, glob='*.json'):
+            paths.append(self.get_truth(path))
+        self.truth_remote = os.path.join(self._inputs_root, self.env, truth_pool_path)
+
+        return paths